Add utm conversions date range
This implements polling via the frontend under the assumption that
most of the time users will be exploring a lot of ranges that overlap
strongly, and hence they will tend to explore more unique ranges than
unique dates. Hence we'd prefer to cache the dates and let the client
build the ranges.

This is also less error-prone than trying to cache ranges, and the least
viable way to implement this would be to have the backend return the
summed values without any caching.