Refactor OsehImage
This new, simplified interface lends itself to avoiding repeated network
requests and greatly improved code legibility. The main advantage of
this interface is it moves the logic of determining when images can be
released to the only spot that can know for certain: the request handler
itself. In return for this, clients only need to specify when they no longer
are using their own copy of the image.

For the most part I adjusted calling code to do the trivial thing: create
a new request handler, then use it as before. This leads to the same
performance as before. The main notable exception is in the admin
area where it's more convenient to reuse the handler, which results
in much faster pagination as well!