Refactor single-pass webgl components
This provides the utility hook useAnimationLoop and the
utility component SinglePassWebGLComponent. When
combined, this allows splitting the rendering logic (i.e.,
webgl calls) from the animation logic (i.e., requestAnimationFrame).
Further, most of the time the animation logic can be implemented
the same way, so this also provides the classes TrivialAnimation,
BezierAnimation, and BezierColorAnimation which implement the
standard animations.

These refactorings are used to reimplement the vertical partly filled
rounded rect, used in the numeric and color prompts. This also fixes
the 50% border radius artifacting incidentally (by ensuring we always
consider a point in one of the quadrants by ensuring some quadrants
are end-inclusive). However, as a result of just the refactoring the
VPFRR supports animating any of its props.

Removes the border color prop, which was included before but didn't
do anything

Changes numeric prompt to using full circles again, which was removed
because VPFRR didn't support it