PickEmotion CSS Transitions -> Animation Loop
The end result is different; not better, not worse, just different. The
code is a bit more complicated. The motivating reason for this change
is simple: there's no CSS transitions to use in native, so in order to
keep parity, I need to implement this without them!

This also incorporates another compatibility change: on native, theres
no convenient way to get the rendered size of components until a frame
*after* they successfully render. This doesn't go as far as delaying
checking the size until a frame later, but it does ensure the interface
is compatible with that and moves the size check to be after the
words are already mounted. Hence the passing WriteableValueWithCallbacks
to the Word component to get the size, rather then precomputing.
Same deal with the container size, since that depends on the word sizes