Remove remaining rerenders during pick emotion animation
There is only ever a single rerender, and only if the extended class
pack is shown. For all other cases, nothing uses react state to rerender.
This takes the worst-case render time (when not showing the ECP) from
1.9ms to 0.5ms on my machine, a 74% improvement. It should be much
more dramatic in react native which is >16ms for a react state rerender;
assuming it falls to 1ms (the target), that'll be a ~94% improvement