Performance (#4)
* Use VWC to reduce rerenders

This is an intermediary commit which reduces the number of react
rerenders during the feel animation to 3, primarily by changing
useJourneyShared to never trigger react rerenders. It introduces
some interesting ideas, like RenderGuardedComponent.

Since we still have some rerenders to remove this isn't done, but
I commit here because I'm going to switch to solving this general
case by having Feature's hooks in general never trigger rerenders

* More work reducing rerenders

* Remove unnecessary type (resolved merge wrong)

* Remove remaining rerenders during pick emotion animation

There is only ever a single rerender, and only if the extended class
pack is shown. For all other cases, nothing uses react state to rerender.
This takes the worst-case render time (when not showing the ECP) from
1.9ms to 0.5ms on my machine, a 74% improvement. It should be much
more dramatic in react native which is >16ms for a react state rerender;
assuming it falls to 1ms (the target), that'll be a ~94% improvement

* Add missing hook documentation