Rework interactive prompts to use VWCs
This brings interactive prompts into the latest ecosystem. It
was the first one to try the value with callbacks strategy, but
before the value with callbacks type was built (and associated
helpers). So this is a relatively straightforward set of changes
to use the new interfaces.

The main cause of verbosity is this now allows each hook to
receive all of its components either via react-rerender or via
value with callbacks, rather than before where each one was
specified individually. This makes the hooks more versatile,
though I think there ought to be a way to reduce how verbose
it makes it to call them

This also reduces avoidable requests in useJourneyShared,
especially reloading the audio necessary.

The most complex rework was useJoinLeave surprisingly enough,
because the old strategy with cancelable promises gets quadratically
more complex the more things there are to listen to. I switched it to a
more reactive approach which avoids this complexity scaling.