Update JourneyFeedbackScreen for native compat
This has the positive effect that the gradient change is now animated
but is otherwise the same, except no longer using css transitions.

Also removes an extra rerender in useUnwrappedVWC (and the component
version, RenderGuardedComponent), where we were instantly setting the
value at most one time, which was enough to block callback loops but
makes understanding the effects of it difficult.