Refactor JourneyPostScreen for app
Also broke out toggleFavorited so that logic is all in one place, then
updated HistoryItem, CourseJourneyItem, and Journey to use that
function. CourseJourneyItem hadn't yet been updated to VWCs, so
did that as well.

I suspect when I go to convert favorites to the app I'll be refactoring
it to avoid the rerenders when favoriting, since I think it currently
rerenders every item