Use react-virtuoso for infinite list
This is primarily intended to bring the implementation more inline with
the react-native version. Comes with a couple of quirks (e.g., seeing
loading elements and a scrollbar which isn't particularly accurate) but
has slightly less jank while scrolling on ios. Performance wise this is
definitely worse, but its still good enough when combined with the VWC
improvements skipping all rerenders of the underlying history items