1.5.0
New media player, home screen copy, and transitions

This update features personalized insights on the home screen, which range from curated mindfulness tips and quotes to provoke thoughtful moments to encouragement and acknowledgement for streaks and goals. It integrates an improved media player that enables closed captioning and tap to mute functionality on all journeys and series introductions, as well as tap-to-seek and standardized styling. Last but not least, it adds satisfying transitions between series listings, series introductions, and series details.